network: move assemble_tcp_address to network module
make sure we have network bits in related source/tests.
