+++ src/libhirte/bus/bus.c
@@ -13,22 +13,6 @@
  *** peer dbus ***
  *****************/
 
-char *assemble_tcp_address(const struct sockaddr_in *addr) {
-        if (addr == NULL) {
-                hirte_log_error("Can not assemble an empty address");
-                return NULL;
-        }
-
-        _cleanup_free_ char *host = typesafe_inet_ntop4(addr);
-        char *dbus_addr = NULL;
-        int r = asprintf(&dbus_addr, "tcp:host=%s,port=%d", host, ntohs(addr->sin_port));
-        if (r < 0) {
-                hirte_log_error("Out of memory");
-                return NULL;
-        }
-        return dbus_addr;
-}
-
 static sd_bus *peer_bus_new(sd_event *event, const char *dbus_description) {
         if (event == NULL) {
                 hirte_log_error("Event loop must be initialized");
