+++ src/libhirte/bus/bus.h
@@ -5,7 +5,6 @@
 #include <systemd/sd-bus.h>
 
 
-char *assemble_tcp_address(const struct sockaddr_in *addr);
 sd_bus *peer_bus_open(sd_event *event, const char *dbus_description, const char *dbus_server_addr);
 int peer_bus_close(sd_bus *peer_dbus);
 sd_bus *peer_bus_open_server(sd_event *event, const char *dbus_description, const char *sender, int fd);
