+++ src/libhirte/common/network.c
@@ -70,3 +70,19 @@ char *typesafe_inet_ntop6(const struct sockaddr_in *addr) {
         inet_ntop(AF_INET6, &addr->sin_addr, dst, INET6_ADDRSTRLEN);
         return dst;
 }
+
+char *assemble_tcp_address(const struct sockaddr_in *addr) {
+        if (addr == NULL) {
+                hirte_log_error("Can not assemble an empty address");
+                return NULL;
+        }
+
+        _cleanup_free_ char *host = typesafe_inet_ntop4(addr);
+        char *dbus_addr = NULL;
+        int r = asprintf(&dbus_addr, "tcp:host=%s,port=%d", host, ntohs(addr->sin_port));
+        if (r < 0) {
+                hirte_log_error("Out of memory");
+                return NULL;
+        }
+        return dbus_addr;
+}
