+++ src/libhirte/test/meson.build
@@ -1,4 +1,3 @@
 # SPDX-License-Identifier: GPL-2.0-or-later
 
-subdir('bus')
 subdir('common')
